Move `DefaultValue` to `shared`
